+++ days/11/day-11.hs
@@ -1,4 +1,4 @@
-{-# LANGUAGE LambdaCase #-}
+
 {-# LANGUAGE MultiWayIf #-}
 
 module Main where
@@ -31,22 +31,37 @@ showMonkey m = show $ mID m
 monkeyInLine :: String -> Bool
 monkeyInLine line = "Monkey" `elem` words line
 
--- happyPathParser :: String -> ???
-happyPathParser str = 
+getMonkeyID :: String -> Int
+getMonkeyID line = read $ init (words line !! 1)
+
+startsWithStarting :: String -> Bool
+startsWithStarting line = "Starting" == head (words line)
+
+monkeyStrings :: String -> [[String]]
+monkeyStrings str =
   let lns = lines str
+  in
+    foldl (\ms ln -> (if ln == "" then []:ms else (head ms ++ [ln]):tail ms) ) [[]] lns
+
+-- happyPathParser :: String -> ???
+happyPathParser str =
+  let -- lns = lines str
+      ms = monkeyStrings str
       blankMnky = Monkey (-1) [] (+0) (-1) (-1)
   in
-    foldl (\mnky ln -> if
-              | (monkeyInLine ln) -> blankMnky { mID = 42 }
-              | otherwise -> undefined ) blankMnky lns
+    map ( foldl (\mnky ln -> if
+              | (monkeyInLine ln) -> mnky { mID = getMonkeyID ln }
+              -- | 
+              | otherwise -> mnky ) blankMnky ) ms
 
 main :: IO ()
 main = do
   f <- readFile "input-11-test.txt"
   -- f <- readFile "input-11.txt"
-  
+
   putStrLn $ replicate 42 '-'
 
   let parsedInput = happyPathParser f
 
-  putStrLn $ unlines $ take 3 parsedInput
+  -- putStrLn $ unlines $ map showMonkey $ take 3 parsedInput
+  print "42"